Default Level Setup
-Added floor.
-Added placeholder player with simple material
-Rigged sun system to follow player.  Directional Light system is
centered on player. Attached a "invisible cube" to the light and moved
it 500 (can be adjusted to any distance) units forward from light.
Rotating the light will cause the emiiter to rotate and move, so
raycasts (when coded) will shoot directly at the player and will always
be emitting from the exact direction of the sun.

No code has been applied yet, just wanted to get the base level setup
saved.