Implemented Critical Feature - Raycasting
Implemented Critical Feature Raycasting, we can now detect if player is
in the sunlight. We may need to review collider boxes as these determine
where the ray hits and not the mesh (only drawnback to this method ive
seen so far). But this should be an easy fix of messing with colliders
in the editor.

Stared at my code for about an hour because the console log was tricking
me into thinking it was only detecting the player on the first hit.
Curse you Collapse button on the console!

Verified that raycasting is only throwing "do this" commands when it
hits the player. Raycast emitter follows player and will follow the
rotation of the sun FLAWLESSLY. Best part is no math was used on my part
to get the positioning of the raycast emitter to always be between
player and sun, and always facing player.

Put in a cloud for demonstration purposes, if you fly into the shadow,
the console stops reporting hits, fly out into the sun and the hit logs
will flow. will tweak logic once gameplay programming begins so a
boolean will be set stating wheather player is in light or not (if hit,
they are in light), which will effect speed.

Currently can only move player in editor, but movement code is coming
tomorrow night (basic left right and jump). Momentum will occur next
week when adding obstacles in. 10 hour workdays in front of a computer
are pretty fun... then coming home to do this.